requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain & OpenAI
langchain==0.1.0
langchain-openai==0.0.2
openai==1.8.0

# Deepgram Transcription
deepgram-sdk==3.1.0

# Airtable Database
pyairtable==2.2.1

# HTTP Clients
requests==2.31.0
httpx==0.26.0

# Environment & Validation
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
tenacity==8.2.3
yt-dlp
